flir / maintenance / product / team / wishlist screens moved into componenet folder
