--- frontend/src/components/Flir/FlirHero.js
@@ -1,18 +1,0 @@
-import React from 'react'
-import {Jumbotron} from "reactstrap";
-import HeroDivider from "../HeroDivider";
-import {Container} from "react-bootstrap";
-
-const FlirHero = () => {
-
-//     return (
-//
-//     //         // <Jumbotron className='FlirHero_hero-img global_rich-black-bg '
-//     //     //             //            fluid>
-//     //     //             // </Jumbotron>
-//         <HeroDivider img={'WishListScreenImg'} show={true} />
-//
-//     )
-};
-
-export default FlirHero
